Move template metafunctions to header
We'll need them in one more file too
